add calcExprs.hpp/cpp
